requirements.txt
accelerate==1.6.0
transformers==4.54.1
diffusers[torch]==0.32.1
ftfy==6.3.1
# albumentations==1.3.0
opencv-python==4.10.0.84
einops==0.7.0
bitsandbytes
lion-pytorch==0.2.3
schedulefree==1.4
pytorch-optimizer==3.10.0
prodigy-plus-schedule-free==1.9.2
prodigyopt==1.1.2
tensorboard>=2.17,<3
safetensors==0.5.3
# Optional dataset tag editor stack.
# Gradio 3.x currently pulls NumPy 1.x, which does not provide Windows cp313 wheels.
# Keep it out of the base install so the main GUI can run on bundled Python 3.13.
# altair==4.2.2
# easygui==0.98.3
toml==0.10.2
voluptuous==0.15.2
huggingface-hub==0.34.3
# for Image utils
imagesize==1.4.1
# for T5XXL tokenizer (SD3/FLUX)
sentencepiece==0.2.1
# for ui
rich==14.1.0
pandas>=2,<3
scipy>=1.13,<2
requests>=2.32,<3
pillow>=10,<12
numpy==2.2.6
# gradio>=3.44,<4
fastapi>=0.95,<1
uvicorn>=0.22,<1
wandb>=0.16,<1
httpx>=0.24,<1
packaging>=24,<26
psutil>=5.9,<8
websockets>=12,<16
# extra
open-clip-torch==2.20.0
lycoris-lora==3.2.0.post2
dadaptation==3.1
